# docs/openapi.yaml
openapi: 3.1.0
info:
  title: x402-micro-tollgate
  version: 0.3.3
  summary: Thin x402 payment gateway + MCP server
  description: |
    Self-hosted HTTP gateway that gates routes with HTTP 402 / x402 micropayments
    (default Base USDC via Coinbase CDP + EIP-3009) and exposes an MCP endpoint.

    Payment headers used by this gateway (do not invent alternate names):
    - Request: `PAYMENT-SIGNATURE` (also `payment-signature`), optional `x-demo-payment`,
      optional `x-merchant-id`
    - Response: `PAYMENT-REQUIRED` on unpaid 402; `PAYMENT-RESPONSE` after settle / demo accept

    Free discovery: `/health`, `/x402/discover`, `/.well-known/x402.json`,
    `/.well-known/agent-card.json`, `/llms.txt`, `/openapi.yaml`,
    `/docs/openapi.yaml`, `/merchants`.
    Gated by default under `GATED_PREFIX` (default `/v1`).
  license:
    name: MIT
  contact:
    email: 2767111713@qq.com
servers:
  - url: https://x402-micro-tollgate.onrender.com
    description: Public demo (Render)
  - url: http://127.0.0.1:8402
    description: Local default
...
paths:
...
  /v1/fetch-md:
    get:
      tags: [paid]
      summary: Paid HTML→Markdown fetch demo
      operationId: getV1FetchMd
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
            format: uri
          description: Public http(s) URL to fetch
        - $ref: "#/components/parameters/MerchantQuery"
        - $ref: "#/components/parameters/MerchantHeader"
        - $ref: "#/components/parameters/PaymentSignature"
        - $ref: "#/components/parameters/DemoPayment"
      responses:
        "200":
          description: Paid — Markdown payload
          headers:
            PAYMENT-RESPONSE:
              $ref: "#/components/headers/PaymentResponse"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "402":
          $ref: "#/components/responses/PaymentRequired"
...
  /mcp:
    post:
      tags: [mcp]
      summary: Streamable HTTP MCP transport
      description: |
        Free path for MCP session handshake / server_info. Individual MCP tools
        (e.g. get_quote, proxy_request) may require x402 payment inside the MCP
        payment wrapper. Prefer Cursor/Claude MCP clients over raw curl.
      operationId: postMcp
      responses:
        "200":
          description: MCP JSON-RPC / SSE stream response
        "400":
          description: Malformed MCP request
components:
  parameters:
    MerchantQuery:
      name: merchant
      in: query
      required: false
      schema:
        type: string
      description: Merchant id from the registry (alternative to x-merchant-id)
    MerchantHeader:
      name: x-merchant-id
      in: header
      required: false
      schema:
        type: string
      description: Merchant id header (case-insensitive)
    PaymentSignature:
      name: PAYMENT-SIGNATURE
      in: header
      required: false
      schema:
        type: string
      description: |
        x402 payment proof. Gateway also accepts lowercase `payment-signature`.
        Required to unlock gated routes in live mode.
    DemoPayment:
      name: x-demo-payment
      in: header
      required: false
      schema:
        type: string
        enum: [settled]
      description: Demo-mode only bypass (`settled`). No on-chain settle.
  headers:
    PaymentRequired:
      description: Encoded x402 PaymentRequired payload (Base64 / protocol encoding via @x402/core)
      schema:
        type: string
    PaymentResponse:
      description: Settlement or demo receipt (Base64 JSON in demo mode)
      schema:
        type: string
  responses:
    PaymentRequired:
      description: |
        Unpaid gated request. Protocol offer is in the `PAYMENT-REQUIRED` header.
        JSON body is a small human-readable summary (not the full accepts list).
      headers:
        PAYMENT-REQUIRED:
          $ref: "#/components/headers/PaymentRequired"
        Cache-Control:
          schema:
            type: string
            example: no-store
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PaymentRequiredBody"
  schemas:
...
    PaymentRequiredBody:
      type: object
      required: [error, price, doc, message]
      properties:
        error:
          type: string
          example: Payment Required
        price:
          type: string
          example: "$0.001 USDC"
        doc:
          type: string
          format: uri
        message:
          type: string